// apps/playground/components/Pressable/Pressable.types.ts
export interface IPosition {
  left: number;
  top: number;
}

export interface IRect extends IPosition {
  right: number;
  bottom: number;
}

/**
 * Describes the height and width of a control
 */
export interface IDimensions {
  width: number;
  height: number;
}